+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/composables/useFlow.ts
@@ -33,7 +33,7 @@ export interface AutoLayoutOptions {
 }
 
 export default function useFlow(phase: NodePhase, flowId?: string) {
-  const vueFlowStore = useVueFlow({ id: flowId })
+  const vueFlowStore = useVueFlow(flowId)
   const editorStore = useEditorStore()
 
   const {