+++ packages/entities/entities-key-sets/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.8.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-key-sets@3.8.2...@kong-ui-public/entities-key-sets@3.8.3) (2025-08-11)
+
+**Note:** Version bump only for package @kong-ui-public/entities-key-sets
+
+
+
+
+
 ## [3.8.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-key-sets@3.8.1...@kong-ui-public/entities-key-sets@3.8.2) (2025-08-07)
 
 **Note:** Version bump only for package @kong-ui-public/entities-key-sets